backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
gunicorn>=22.0
sqlalchemy>=2.0
alembic>=1.14
pydantic>=2.10
pydantic-settings>=2.7
python-dotenv>=1.0
anthropic>=0.42
httpx>=0.28
python-multipart>=0.0.20
psycopg2-binary>=2.9
bcrypt>=4.0
passlib[bcrypt]>=1.7
python-jose[cryptography]>=3.3
gspread>=6.0
google-auth>=2.0
